requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
multidict==6.4.3
narwhals==1.37.1
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
pyarrow==20.0.0
pydeck==0.9.1
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
openpyxl==3.1.5
dotenv==0.9.9
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
fastapi>=0.104.1,<0.111.0
uvicorn[standard]>=0.23.2,<0.25.0
pydantic>=1.10.12,<2.6.0
python-multipart>=0.0.6,<0.0.8
googleapis-common-protos==1.70.0
filetype==1.2.0